miyog/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings
python-multipart
celery==5.3.4
redis==5.0.1
moviepy==1.0.3
requests==2.31.0
google-generativeai>=0.5.0
openai-whisper
mutagen
Pillow==9.5.0
python-jose[cryptography]
httpx
sqlalchemy
kokoro>=0.8.2
soundfile
numpy
torch
scipy
psycopg2-binary
boto3